rsync the kernel_target with --copy-links
For the other two rsync instance, these folders are created as
installable (i.e. the symlinks in them make sense to copy around).

If someone does a:

    install-kernel.sh -v arch/x86_64/boot/bzImage 10.10.224.13

for example, that points to a symlink. In this case we're literally
copying the file, we don't want to replicate the link.. so don't.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>