Add a -p <port> option
rsync/scp is used here heavily, and port 22 is the default.

If you're trying to copy to a machine that has a different port by
default (say qemu instance with SLIRP going and port forwarding) you'll
want a different port than the default.

Add that as a parameter.

Example of when I'd use it is when running qemu like below:

    % sudo ./automotive-image-runner --verbose --memory 4G /tmp/CentOS-Stream-GenericCloud-x86_64-9-latest.x86_64.qcow2
    port: 2222 → 22
    MAC: FE:b6:00:4b:d1:7a
    Image: /tmp/CentOS-Stream-GenericCloud-x86_64-9-latest.x86_64.qcow2
    Running: /usr/bin/qemu-system-x86_64 -drive file=/usr/share/OVMF/OVMF_CODE.fd,if=pflash,format=raw,unit=0,readonly=on -drive file=/usr/share/OVMF/OVMF_VARS.fd,if=pflash,format=raw,unit=1,snapshot=on,readonly=off -smp 20 -enable-kvm -m 4G -machine q35 -cpu host -device virtio-net-pci,netdev=n0,mac=FE:b6:00:4b:d1:7a -netdev user,id=n0,net=10.0.2.0/24,hostfwd=tcp::2222-:22 -drive file=/tmp/CentOS-Stream-GenericCloud-x86_64-9-latest.x86_64.qcow2,index=0,media=disk,format=qcow2,if=virtio,id=rootdisk,snapshot=off

Then I can use it like:

    % ~/git/install-kernel-script/install-kernel.sh -p 2222 arch/x86_64/boot/bzImage 127.0.0.1

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>